% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{bibentry-class}
\alias{bibentry-class}
\title{The literature module}
\description{
The eml-literature module contains information that describes literature
resources. It is intended to provide overview information about the
literature citation, including title, abstract, keywords, and contacts.
Citation types follow the conventions laid out by EndNote, and there is an
attempt to represent a compatible subset of the EndNote citation types.
These citation types include: article, book, chapter, edited book,
manuscript, report, thesis, conference proceedings, personal communication,
map, generic, audio visual, and presentation. The "generic" citation type
would be used when one of the other types will not work (cite eml
documentation)).
The module includes
      eml-documentation,
      eml-resource,
      eml-coverage,
      eml-party,
      eml-access,
      eml-project
Inclusion of superclass
}

