% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_attributes.R
\name{build_factors}
\alias{build_factors}
\title{build factor table}
\usage{
build_factors(att_table, data)
}
\arguments{
\item{att_table}{(data.frame) input attributes table}

\item{data}{(data.frame) input data}
}
\description{
builds factor table for shiny app
}
