% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_attributes.R
\name{build_units_table}
\alias{build_units_table}
\title{build units table}
\usage{
build_units_table(in_units, eml_units)
}
\arguments{
\item{in_units}{input units}

\item{eml_units}{eml units}
}
\description{
builds unit table for shiny app
}
