\docType{methods}
\name{c,codeDefinition-method}
\alias{c,attribute,ANY-method}
\alias{c,attribute-method}
\alias{c,codeDefinition,ANY-method}
\alias{c,codeDefinition-method}
\title{concatenate attribute}
\usage{
\S4method{c}{codeDefinition}(x, ...)

\S4method{c}{attribute}(x, ...)
}
\arguments{
  \item{...}{objects to be concatenated.}
}
\description{
concatenate attribute
}

