\name{citation.eml}
\alias{citation.eml}
\title{Return the citation information to an EML file.}
\usage{
  citation.eml(doc)
}
\arguments{
  \item{doc}{an object of the class eml, as returned from
  `eml_read`, or the filename of an eml object}
}
\description{
  Return the citation information to an EML file.
}
\details{
  Can simply be called as `citation(doc)` after reading in
  a file with eml_read.
}
\seealso{
  \code{\link{citation}}
}

