\docType{methods}
\name{c,citation-method}
\alias{c,citation-method}
\title{concatenate}
\usage{
\S4method{c}{citation}(x, ..., recursive = FALSE)
}
\arguments{
  \item{x,...}{citations to concatenate}

  \item{recursive}{Needed for compatibility with generic,
  otherwise ignored}
}
\description{
concatenate
}

