% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_classes.R
\docType{methods}
\name{c,dataTable-method}
\alias{c,dataTable-method}
\title{concatenate}
\usage{
\S4method{c}{dataTable}(x, ..., recursive = FALSE)
}
\arguments{
\item{x, ...}{dataTables to concatenate}

\item{recursive}{Needed for compatibility with generic, otherwise ignored}
}
\description{
concatenate
}

