\docType{methods}
\name{c,institution-method}
\alias{c,institution-method}
\title{concatenate}
\usage{
\S4method{c}{institution}(x, ..., recursive = FALSE)
}
\arguments{
  \item{x,...}{institutions to concatenate}

  \item{recursive}{Needed for compatibility with generic,
  otherwise ignored}
}
\description{
concatenate
}

