% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_classes.R
\docType{methods}
\name{c,otherEntity-method}
\alias{c,otherEntity-method}
\title{concatenate}
\usage{
\S4method{c}{otherEntity}(x, ..., recursive = FALSE)
}
\arguments{
\item{x, ...}{otherEntities to concatenate}

\item{recursive}{Needed for compatibility with generic, otherwise ignored}
}
\description{
concatenate
}

