% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/party_classes.R
\docType{methods}
\name{c,performer-method}
\alias{c,performer-method}
\title{concatenate}
\usage{
\S4method{c}{performer}(x, ..., recursive = FALSE)
}
\arguments{
\item{x, ...}{performers to concatenate}

\item{recursive}{Needed for compatibility with generic, otherwise ignored}
}
\description{
concatenate
}

