\docType{methods}
\name{c,proceduralStep-method}
\alias{c,proceduralStep-method}
\title{concatenate}
\usage{
\S4method{c}{proceduralStep}(x, ..., recursive = FALSE)
}
\arguments{
  \item{x,...}{proceduralSteps to concatenate}

  \item{recursive}{Needed for compatibility with generic,
  otherwise ignored}
}
\description{
concatenate
}

