% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_classes.R
\docType{methods}
\name{c,qualityControl-method}
\alias{c,qualityControl-method}
\title{concatenate}
\usage{
\S4method{c}{qualityControl}(x, ..., recursive = FALSE)
}
\arguments{
\item{x, ...}{qualityControls to concatenate}

\item{recursive}{Needed for compatibility with generic, otherwise ignored}
}
\description{
concatenate
}

