\docType{methods}
\name{c,recipient-method}
\alias{c,recipient-method}
\title{concatenate}
\usage{
\S4method{c}{recipient}(x, ..., recursive = FALSE)
}
\arguments{
  \item{x,...}{recipients to concatenate}

  \item{recursive}{Needed for compatibility with generic,
  otherwise ignored}
}
\description{
concatenate
}

