% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml_unit_classes.R
\docType{methods}
\name{c,stmml_unitType-method}
\alias{c,stmml_unitType-method}
\title{concatenate}
\usage{
\S4method{c}{stmml_unitType}(x, ..., recursive = FALSE)
}
\arguments{
\item{x, ...}{objects to concatenate}

\item{recursive}{Needed for compatibility with generic, otherwise ignored}
}
\description{
concatenate
}

