% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_classes.R
\docType{methods}
\name{c,taxonomicClassification-method}
\alias{c,taxonomicClassification-method}
\title{concatenate}
\usage{
\S4method{c}{taxonomicClassification}(x, ..., recursive = FALSE)
}
\arguments{
\item{x, ...}{taxonomicClassifications to concatenate}

\item{recursive}{Needed for compatibility with generic, otherwise ignored}
}
\description{
concatenate
}

