% Generated by roxygen2 (4.0.1): do not edit by hand
\name{create_custom_unit}
\alias{create_custom_unit}
\title{create a custom unit}
\usage{
create_custom_unit(id, unitType = NULL, parentSI = NULL,
  multiplierToSI = NULL, name = id, constantToSI = NULL,
  abbreviation = NULL, description = NULL)
}
\arguments{
\item{id}{The id of the unit, given in camelCase with `Per` signifying division, Squared for second power, etc.}

\item{name}{The name of the unit, by default the same as the id.}

\item{unitType}{The type of the unit.}

\item{parentSI}{the parent SI unit from the Standard Units list}

\item{multiplierToSI}{The multiplicative factor to convert the given unit into the parent unit}

\item{constantToSI}{An additive constant needed for the conversion, such as farenheight to celsius}

\item{abbreviation}{Optional abbreviation}

\item{description}{a plain text description of the unit}
}
\value{
The unit definition (invisibly), as a list object.  A list of these returned objects
can be passed to the \code{\link{eml}} function directly to define additional units, bypassing
the config mechanism used by default.
}
\description{
create a custom unit
}
\details{
Defines a custom unit whose id is used in the unit definition metadata. See https://github.com/ropensci/EML/issues/12
}
\examples{
create_custom_unit(id = "metersSquaredPerHectare",
                    parentSI = "dimensionless",
                    unitType = "dimensionless",
                    multiplierToSI = "0.0001",
                    description = "Square meters per hectare")
}

