\docType{data}
\name{dat}
\alias{dat}
\title{A small example data.set object, consisting of data.frame and unit and column metadata}
\description{
A small example data.set object, consisting of data.frame
and unit and column metadata
}
\author{
Carl Boettiger \email{cboettig@ropensci.org}
}
\keyword{data}

