% Generated by roxygen2 (4.0.2): do not edit by hand
\name{detect_class}
\alias{detect_class}
\title{Helper function to attach classes to metadata}
\usage{
detect_class(meta, col.classes)
}
\arguments{
\item{meta}{a list containing attribute-level
metadata for each column.  see details.}

\item{col.classes}{a character vector of column classes from the data.}
}
\description{
Helper function to attach classes to metadata
}
\details{
each element of meta list corresponds to a consecutive
column, and gives a list containing the column name (as appearing in the
data-frame), the column description (e.g. from col.defs) and the unit defs
(e.g. from unit.defs) as  a (named) character vector.
}

