% Generated by roxygen2 (4.0.1): do not edit by hand
\name{eml_attributeList}
\alias{eml_attributeList}
\title{eml_attributeList}
\usage{
eml_attributeList(dat = NULL, meta = NULL, col.classes = sapply(dat,
  class), col.defs = NULL, unit.defs = NULL)
}
\arguments{
\item{dat}{a data.frame (or a data.set which has col.defs and unit.defs)}

\item{meta}{a list of column and unit metadata, OR}

\item{col.defs}{the column definitions (if not given in meta)}

\item{unit.defs}{the unit definitions (if not given in meta)}

\item{col.classes}{the column classes of dat, if dat not given (otherwise inferred from dat)}
}
\description{
eml_attributeList
}

