% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml_to_jsonld.R
\name{eml_boundingCoordinates_to_GeoShape}
\alias{eml_boundingCoordinates_to_GeoShape}
\title{Convert an EML boundingCoordinates to a list suitable for conversion to
Schema.org JSON-LD#'}
\usage{
eml_boundingCoordinates_to_GeoShape(boundingCoordinates)
}
\arguments{
\item{boundingCoordinates}{An EML boundingCoordinate object}
}
\value{
(list) A list of statements for insertion into the JSON-LD document
The first point is the lower corner, the second point is the upper corner.
Appears to be in format lat lon lat lon
}
\description{
Convert an EML boundingCoordinates to a list suitable for conversion to
Schema.org JSON-LD#'
}
