% Generated by roxygen2 (4.0.1): do not edit by hand
\name{eml_dataTable}
\alias{eml_dataTable}
\title{constructor helper function for data table}
\usage{
eml_dataTable(dat, name = "data table", description = character(0),
  filename = generate_filename(name), col.defs = NULL, unit.defs = NULL,
  col.classes = NULL, additionalInfo = character(0), ...)
}
\arguments{
\item{dat}{a data.set object, data.frame object}

\item{name}{the the name of the dataTable}

\item{description}{a description of the dataTable}

\item{filename}{the filename of the csv file created}

\item{...}{additional arguments (not yet implemented)}

\item{col.defs}{Natural language definitions of each column. Should be a character
vector of length equal to each column, defined in the same order as the columns are given.}

\item{unit.defs}{A list of length equal to the number of columns defining the units for each
column. See examples.}

\item{col.classes}{column classes, primarily for use if dat is
a path to a csv rather than a native R object. Optional otherwise.}

\item{additionalInfo}{character vector of any additional metadata lacking a defined slot.}
}
\value{
a dataTable element
}
\description{
constructor helper function for data table
}

