\name{eml_distribution}
\alias{eml_distribution}
\title{Create an eml_distribution node}
\usage{
  eml_distribution(url, node = NULL)
}
\arguments{
  \item{url}{the online url at which a resource can be
  found}
}
\value{
  an EML distribution node
}
\description{
  Defines how a resource is distributed
}
\details{
  Currently only provides online distribution types FIXME
  should be extended to support offline and inline
  distribution
}

