% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml_find.R
\name{eml_find}
\alias{eml_find}
\title{eml_find}
\usage{
eml_find(node, element)
}
\arguments{
\item{node}{an EML S4 object; e.g. `eml` or any of it's child/descendant slots.}

\item{element}{the name of an EML element to find}
}
\value{
the requested element as an EML S4 object
}
\description{
find all occurrences of an EML element in at a given node at any depth
}

