% Generated by roxygen2 (4.0.1): do not edit by hand
\name{eml_geographicCoverage}
\alias{eml_geographicCoverage}
\title{Helper constructor for creating geographicCoverage node}
\usage{
eml_geographicCoverage(description, NSEWbox = NULL)
}
\arguments{
\item{description}{the text-based description of the geography.  This can also
be an`eml::geographicCoverage` object.}

\item{NSEWbox}{the North, South, East and West coordinates, in signed decimal degrees
latitude/longitude, respectively.  Additionally, the minimum and maximum altitude in
meters can optionally be given as the 5th and 6th values of this numeric vector.  This
value can only be left as NULL if description is a completed `eml::geographicCoverage`
object.}
}
\description{
Implements only a subset of possible specifications for coverage
}

