% Generated by roxygen2 (4.0.1): do not edit by hand
\name{eml_get}
\alias{eml_get}
\title{Extract content from an EML object}
\usage{
eml_get(eml, x = c("id", "version", "csv_filepaths", "unit.defs", "col.defs",
  "citation_info", "keywords", "coverage", "contact", "creator",
  "attributeList", "data.set"))
}
\arguments{
\item{eml}{an EML object}

\item{x}{the desired content we should extract from the eml object}
}
\value{
the requested content.  Format will depend on the request.
}
\description{
Extract content from an EML object
}
\details{
All content can always be accessed using the standard S4 subsetting mechanisms.
This function merely provides a convenient interface for returning more commonly requested
content in forms that may be more convenient than the S4 structure.  These formats mirror the
input formats used by the corresponding helper constructors.
}

