% Generated by roxygen2 (4.0.1): do not edit by hand
\name{eml_get_config}
\alias{eml_get_config}
\title{eml get config}
\usage{
eml_get_config(x = ls(envir = EMLConfig))
}
\arguments{
\item{x}{name of the configuration option desired (gives all by default)}
}
\value{
display the current configuration options
}
\description{
eml get config
}

