% Generated by roxygen2 (4.0.2): do not edit by hand
\name{eml_knb}
\alias{eml_knb}
\title{Upload EML to knb}
\usage{
eml_knb(file, permanent = FALSE, public = TRUE,
  mn_nodeid = "urn:node:KNB", cli = D1Client("PROD", mn_nodeid))
}
\arguments{
\item{file}{the filename of the EML file to upload}

\item{permanent}{logical, indicate whether we should use the production server.
default is FALSE, implies use of the Demo server, overriding any arguments set
to mn_nodeid and cli.}

\item{public}{logical, make object public?  Otherwise only authorized users may access file.}

\item{mn_nodeid}{The server ID to which we upload (if permanent is FALSE).  The default
path is to the KNB, which can index all EML metadata through metacat.}

\item{cli}{the dataone client object. See dataone::D1Client}
}
\value{
the EML identifier, csv file identifier, and dataone package identifier
 (a metadata object associating the two files in the dataone repository).
}
\description{
Upload EML to knb
}
\details{
knb requires authentication.
}

