% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml_physical.R
\name{eml_physical}
\alias{eml_physical}
\title{eml_physical constructor}
\usage{
eml_physical(dat, filename = character(),
  characterEncoding = guess_encoding(), ...)
}
\arguments{
\item{dat}{a data.frame or path to existing csv file}

\item{filename}{the filename to be used. Ignored if `dat` 
is a path to a file.  Can include the full path (e.g. if
file is to be written to a different directory).}

\item{characterEncoding}{string defining the encoding used. will guess based on R's configuration (sensible for data.frames or csv files produced by R, but not otherwise).}

\item{...}{additional arguments passed to eml_dataFormat, eml_distribution or write.csv.}
}
\description{
eml_physical constructor
}

