\name{eml_R_software}
\alias{eml_R_software}
\title{Create a software node given the name of an R package}
\usage{
  eml_R_software(packagename)
}
\arguments{
  \item{packagename}{the name of the R package for which we
  want an EML documentation node See issue
  [#32](https://github.com/ropensci/reml/issues/32)}
}
\description{
  Create a software node given the name of an R package
}

