\name{eml_software}
\alias{eml_software}
\title{Create  an EML Software node}
\usage{
  eml_software(license, version, .implementation,
    .dependency = list(), ..., node = NULL)
}
\arguments{
  \item{license}{Text string containing the software
  license}

  \item{version}{of the software}

  \item{node}{an existing software node we wish to append
  to}
}
\description{
  Create an EML Software node
}

