% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml_to_jsonld.R
\name{eml_temporalCoverage_to_Text}
\alias{eml_temporalCoverage_to_Text}
\title{Convert an EML temporalCoverage to a list suitable for conversion to
Schema.org JSON-LD}
\usage{
eml_temporalCoverage_to_Text(temporalCoverage)
}
\arguments{
\item{temporalCoverage}{An EML temporalCoverage object}
}
\value{
(list) A list of statements for insertion into the JSON-LD document
}
\description{
Convert an EML temporalCoverage to a list suitable for conversion to
Schema.org JSON-LD
}
