% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_TextType.R
\name{get_abstract}
\alias{get_abstract}
\title{Get the text content of the abstract}
\usage{
get_abstract(node, as = "character")
}
\arguments{
\item{node}{An EML object or child/descendant object}

\item{as}{The desired return type, see details (optional)}
}
\value{
A text represenation of the abstract
}
\description{
Gets a base R representation of the abstract in the given document, such as a
character vector.
}
\details{
This function uses Pandoc to convert the XML content contained in the
abstract into Markdown and then into a base R type such a character vector or
list. This is a particularly useful function to use if you want to work with
the text content of your abstract to perform downstream processing. A special
function like this one is required because the abstract element in EML is a
TexType which supports a docbook-like structure including paragraphs,
sections, titles, and much more. To work around this, the abstract XML is
converted into Markdown and then read back in as a character vector and,
optionally, converted to your desired type.

A special case is impelemented for character vectors in that a character
vector of length one is always returned.
}
\examples{
f <- system.file("examples/hf205.xml", package = "EML")
eml <- read_eml(f)
# As a single character vector
get_abstract(eml, as = "character")
# As a list of lines
get_abstract(eml, as = "list")
}

