% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coverage.R
\name{get_taxonomicCoverage}
\alias{get_taxonomicCoverage}
\title{get_taxanomicCoverage}
\usage{
get_taxonomicCoverage(taxonomicCoverage)
}
\arguments{
\item{taxonomicCoverage}{a taxonomicCoverage EML element}
}
\value{
a data frame with column for rank and column for value, see examples
}
\description{
get_taxanomicCoverage
}
\examples{
f <- system.file("xsd/test/eml-i18n.xml", package="EML")
eml <- read_eml(f)
df <- EML:::get_taxonomicCoverage(eml@dataset@coverage@taxonomicCoverage)
}

