% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_unitList.R
\name{get_unitList}
\alias{get_unitList}
\title{get_unitList}
\usage{
get_unitList(unitList)
}
\arguments{
\item{unitList}{a unitList object.  Usually found in eml@additionalMetadata[[1]]@metadata}
}
\value{
a list with two data.frames: "units", a table defining unit names, types, and conversions to SI,
and "unitTypes", defining the type of unit. For instance, the unit table could define "Hertz" as a unit
of unitType frequency, and the unitType define frequency as a type whose dimension is 1/time.
}
\description{
get_unitList
}
\examples{
## Read in the definitions of standard units:
f <- system.file("xsd/eml-unitDictionary.xml", package = "EML")
eml <- read_eml(f)
get_unitList(eml)


## note the eml-unitDictionary.xml just defines 
}

