% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_attributes.R
\name{htmlwidgets_attributes}
\alias{htmlwidgets_attributes}
\title{Launch attributes htmlwidget}
\usage{
htmlwidgets_attributes(df, type = NULL)
}
\arguments{
\item{df}{(data.frame) the data.frame of data that needs an attribute table}

\item{type}{(character) either "attributes", "units", or "factors"}
}
\description{
Used to call handsontable html widget to build attributes
}
