% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_attributes.R
\name{is_standardUnit}
\alias{is_standardUnit}
\title{is_standardUnit}
\usage{
is_standardUnit(x)
}
\arguments{
\item{x}{name of unit to check}
}
\value{
TRUE if unit is exact match to the id of
 a unit in the Standard Units Table, FALSE otherwise.
}
\description{
is_standardUnit
}
\examples{
is_standardUnit("amperePerMeter") # TRUE
is_standardUnit("speciesPerSquareMeter") # FALSE
}
