\name{match_unit}
\alias{match_unit}
\title{Helper function to match the unit against standard unit definition
Documentation: http://knb.ecoinformatics.org/software/eml/eml-2.1.1/eml-unitTypeDefinitions.html
XML discription:}
\usage{
  match_unit(char)
}
\description{
  Helper function to match the unit against standard unit
  definition Documentation:
  http://knb.ecoinformatics.org/software/eml/eml-2.1.1/eml-unitTypeDefinitions.html
  XML discription:
}
\keyword{internal}

