% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml_validate.R
\name{message_validation_error}
\alias{message_validation_error}
\title{message_validation_error}
\usage{
message_validation_error(error)
}
\arguments{
\item{error}{The validation error (XML::XMLError)}
}
\value{
Nothing.
}
\description{
Create a useful message() for an XML validation error.
}

