\name{metadata_wizard}
\alias{metadata_wizard}
\title{Generate unit and column metadata for a data.frame}
\usage{
metadata_wizard(dataframe)
}
\arguments{
  \item{dataframe}{a data.frame object}
}
\value{
a metadata list object, such as could be passed to
eml_dataTable
}
\description{
Utility to help generate unit and column metadata for a
data.frame
}

