% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_TextType.R
\name{node_to_file}
\alias{node_to_file}
\title{Converts a node into a file by running it through Pandoc}
\usage{
node_to_file(node, to = "html", output = tempfile(class(node), fileext =
  paste0(".", to)))
}
\arguments{
\item{node}{any node that supports a docbook-like structure (abstract, methods, etc)}

\item{to}{desired format, default is html, but can be any type supported by pandoc (docx, md, etc)}

\item{output}{name of the desired output file}
}
\value{
creates a file requested.
}
\description{
Converts a node into a file by running it through Pandoc
}

