% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_person.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{as_emld}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{emld}{\code{\link[emld]{as_emld}}}
}}

