% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml_to_jsonld.R
\name{remove_extra_empty_lines}
\alias{remove_extra_empty_lines}
\title{Remove leading, duplicate interior, and trailing empty strings}
\usage{
remove_extra_empty_lines(lines)
}
\arguments{
\item{lines}{(character) A character vector of the lines of text to modify}
}
\value{
(character)
}
\description{
This function is a workaround for the odd behavior Pandoc exhibits when
converting from DocBook to plain (text) where leading, duplicate interior,
and trailing empty lines are included in the output.
}
