% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_party.R
\name{set_associatedParty}
\alias{set_associatedParty}
\title{Create an associatedParty object}
\usage{
set_associatedParty(givenNames, surName, organizationNames, positionNames,
  addresses, phones, electronicMailAddresses, onlineUrls, userIds, salutations,
  role)
}
\arguments{
\item{givenNames}{(character) One or more given names for the party's individualName}

\item{surName}{(character) The sur name of the party's individualName}

\item{organizationNames}{(character) One or more organization names for the party}

\item{positionNames}{(character) One or more position names for the party}

\item{addresses}{(address) One or more addresses for the party}

\item{phones}{(character|phone) One or more phones for the party}

\item{electronicMailAddresses}{(character) One or more email addresses for the party}

\item{onlineUrls}{(character) One or more online URLs for the party}

\item{userIds}{(userId) One or more user IDs for the party}

\item{salutations}{(character) One or more salutations for the party's individualName}

\item{role}{(character) A role for the party}
}
\value{
(associatedParty) An associatedParty object
}
\description{
Create an associatedParty object
}
\examples{
# A simple example
set_associatedParty("Bryce", "Mecum", role = "contributor")
# A more complete example
}
