% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_attributes.R
\name{set_attributes}
\alias{set_attributes}
\title{set_attributes}
\usage{
set_attributes(attributes, factors)
}
\arguments{
\item{attributes}{a joined table of all attribute metdata
(such as returned by \code{\link{get_attributes}}, see details)}

\item{factors}{a table with factor code-definition pairs; see details}
}
\value{
an eml "attributeList" object
}
\description{
set_attributes
}

