% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_party.R
\name{set_contact}
\alias{set_contact}
\title{Create a contact object}
\usage{
set_contact(...)
}
\arguments{
\item{givenNames}{(character) One or more given names for the party's individualName}

\item{surName}{(character) The sur name of the party's individualName}

\item{organizationNames}{(character) One or more organization names for the party}

\item{positionNames}{(character) One or more position names for the party}

\item{addresses}{(address) One or more addresses for the party}

\item{phones}{(character|phone) One or more phones for the party}

\item{electronicMailAddresses}{(character) One or more email addresses for the party}

\item{onlineUrls}{(character) One or more online URLs for the party}

\item{userIds}{(userId) One or more user IDs for the party}

\item{salutations}{(character) One or more salutations for the party's individualName}
}
\value{
(contact) A contact object
}
\description{
Create a contact object
}
\examples{
set_contact("Mary", "Scientist")
}
