% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_party.R
\name{set_phone}
\alias{set_phone}
\title{Create a phone object}
\usage{
set_phone(number, phonetype)
}
\arguments{
\item{number}{(character) The phone number}

\item{phonetype}{(character) The type of phone number}
}
\value{
(phone) A phone object
}
\description{
Create a phone object
}
\examples{
set_phone("1-555-555-5555")
set_phone("1-555-555-5556", "facsimilie")
}
