% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_section.R
\name{set_section}
\alias{set_section}
\title{set_section}
\usage{
set_section(file = NULL)
}
\arguments{
\item{file}{path to any document recognized by pandoc, such as Word .docx file or markdown file.}
}
\value{
a ListOfsection object that can be used in constructing any element with section slots, see examples
}
\description{
For any EML element that takes a "section" argument, this function can be used to generate the appropriate EML from a markdown-formatted file.
}
\details{
See http://rmarkdown.rstudio.com and http://pandoc.org for more details about coversions with pandoc & rmarkdown.  
Pandoc comes pre-installed in RStudio
}
\examples{
f <- system.file("examples/hf205-methods.md", package = "EML")
set_section(f)
new("description", section = set_section(f))

## Can also import from methods written in a .docx MS Word file. 
f <- system.file("examples/hf205-methods.docx", package = "EML")
set_section(f)
new("description", section = set_section(f))

}

