% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_software.R
\name{set_software}
\alias{set_software}
\title{set_software}
\usage{
set_software(codemeta)
}
\arguments{
\item{codemeta}{codemeta object, see examples}
}
\value{
an eml software element
}
\description{
set_software
}
\examples{
cm <- jsonlite::read_json(system.file("extdata/codemeta.json", package = "EML"))
software <- set_software(cm)
my_eml <- eml$eml(packageId = "eml-1.2", system = "knb", software = software)

# write_eml(my_eml, "test.xml")
}
