% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_unitList.R
\name{set_unitList}
\alias{set_unitList}
\title{set_unitList}
\usage{
set_unitList(units, unitTypes = NULL)
}
\arguments{
\item{units}{a data.frame describing the custom units, see details.}

\item{unitTypes}{optional, a data.frame defining any additional unitTypes not already defined}
}
\value{
a stmml:unitList element that can be included in additionalMetadata
}
\description{
set_unitList
}
\details{
The units data.frame must have the following columns:
 - name: the name of unit (singular). e.g. 'meter', 'second'
 - unitType: the base type of unit, e.g. 'length'.  If not from a standard type, a new unitType must be provided
 - multiplierToSI: the multiplicative constant to convert to the SI unit. 
 - parentSI: the name of the parent SI unit, e.g. second.  
 - description: a text string describing the unit of measure.
 The following columns are optional:
 - id: usually the same as the name of the unit, e.g. second
 - abbreviation: common abbreviation, e.g. s
 - constantToSI: an additive constant to convert to the equivalent SI unit. If not given, default is "0"

The unitType table must have the following columns:
 - name:
 - id: optional, default is same as the name
 - dimension: name of a base dimension of the unit
 - power: the power to which the dimension is raised (NA implies power of 1)
 
 For instance, the following table defines three base unitTypes: 
 speed (i.e. m/s), acceleration (m/s^2), and frequency (1/s)
 
 name,         dimension, power,
 speed,        length,    NA,
 speed,        time,      "-1",
 acceleration, length,    NA,
 acceleration, time,      "-2",
 frequency,    time,      "-1",
}

