% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_party.R
\name{set_userId}
\alias{set_userId}
\title{Create a \code{userId} object}
\usage{
set_userId(userId, directory)
}
\arguments{
\item{userId}{(character) The user ID}

\item{directory}{(character) The directory under which the user ID applies}
}
\value{
(userId) A userId object
}
\description{
Create a \code{userId} object
}
\examples{
set_userId("alice", "ldap://ldap.example.org")
set_userId("https://orcid.org/0000-0002-0381-3766", "https://orcid.org")
}
