% Generated by roxygen2 (4.0.1): do not edit by hand
\name{show_yaml}
\alias{show_yaml}
\title{Convert EML into a YAML file}
\usage{
show_yaml(object)
}
\arguments{
\item{object}{any object defined by EML that can be coerced into an XML format}
}
\value{
a yaml representation of the object, also printed to screen.
}
\description{
Convert EML into a YAML file
}

