% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_methods.R
\name{species}
\alias{species}
\title{get species information}
\usage{
species(taxonomicCoverage)
}
\arguments{
\item{taxonomicCoverage}{an object from which taxonomic coverage information might be extracted}
}
\value{
species names as a character string
}
\description{
get species information
}

