% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_attributes.R
\name{table_to_r}
\alias{table_to_r}
\title{handsontable to r}
\usage{
table_to_r(table)
}
\arguments{
\item{table}{input table}
}
\description{
Takes a handsontable and converts to r data.frame for shiny app
}
