% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml.R
\name{write_eml}
\alias{write_eml}
\title{write_eml}
\usage{
write_eml(eml, file = NULL, ...)
}
\arguments{
\item{eml}{an eml class object}

\item{file}{file name to write XML. If NULL, will return XML as character string.}

\item{...}{additional arguments to \code{\link{saveXML}}}
}
\value{
If file is not specified, the result is a character string containing
   the resulting XML content. Otherwise return silently.
}
\description{
write_eml
}
\examples{
f <- system.file("xsd/test", "eml.xml", package = "eml2")
eml <- read_eml(f)
write_eml(eml)
}

