% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_to_xml.R
\name{xml_to_s4}
\alias{xml_to_s4}
\title{root <- xml2::xml_new_root("eml", "xmlns:eml" = "eml://ecoinformatics.org/eml-2.1.1")
eml <- EML::read_eml("inst/examples/example-eml-2.1.1.xml")
xml <- s4_to_xml(eml@dataset, root)
xml2::write_xml(xml, "test.xml")}
\usage{
xml_to_s4()
}
\description{
root <- xml2::xml_new_root("eml", "xmlns:eml" = "eml://ecoinformatics.org/eml-2.1.1")
eml <- EML::read_eml("inst/examples/example-eml-2.1.1.xml")
xml <- s4_to_xml(eml@dataset, root)
xml2::write_xml(xml, "test.xml")
}
