\docType{methods}
\name{c,additionalMetadata-method}
\alias{c,additionalMetadata-method}
\title{concatenate}
\usage{
\S4method{c}{additionalMetadata}(x, ..., recursive = FALSE)
}
\arguments{
  \item{x,...}{additionalMetadatas to concatenate}

  \item{recursive}{Needed for compatibility with generic,
  otherwise ignored}
}
\description{
concatenate
}

