% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/literature_classes.R
\docType{methods}
\name{c,Citation-method}
\alias{c,Citation-method}
\title{concatenate}
\usage{
\S4method{c}{Citation}(x, ..., recursive = FALSE)
}
\arguments{
\item{x, ...}{citations to concatenate}

\item{recursive}{Needed for compatibility with generic, otherwise ignored}
}
\description{
concatenate
}

