\docType{methods}
\name{c,codeDefinition-method}
\alias{c,codeDefinition-method}
\title{concatenate}
\usage{
\S4method{c}{codeDefinition}(x, ..., recursive = FALSE)
}
\arguments{
  \item{x,...}{codeDefinitions to concatenate}

  \item{recursive}{Needed for compatibility with generic,
  otherwise ignored}
}
\description{
concatenate
}

