\docType{methods}
\name{c,creator-method}
\alias{c,creator-method}
\title{concatenate}
\usage{
\S4method{c}{creator}(x, ..., recursive = FALSE)
}
\arguments{
  \item{x,...}{creators to concatenate}

  \item{recursive}{Needed for compatibility with generic,
  otherwise ignored}
}
\description{
concatenate
}

