\docType{methods}
\name{c,metadataProvider-method}
\alias{c,metadataProvider-method}
\title{concatenate}
\usage{
\S4method{c}{metadataProvider}(x, ..., recursive = FALSE)
}
\arguments{
  \item{x,...}{metadataProviders to concatenate}

  \item{recursive}{Needed for compatibility with generic,
  otherwise ignored}
}
\description{
concatenate
}

