% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_classes.R
\docType{methods}
\name{c,methodStep-method}
\alias{c,methodStep-method}
\title{concatenate}
\usage{
\S4method{c}{methodStep}(x, ..., recursive = FALSE)
}
\arguments{
\item{x, ...}{methodSteps to concatenate}

\item{recursive}{Needed for compatibility with generic, otherwise ignored}
}
\description{
concatenate
}

