% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{c,methodStep-method}
\alias{c,methodStep-method}
\title{concatenate}
\usage{
\S4method{c}{methodStep}(x, ..., recursive = FALSE)
}
\arguments{
\item{x,...}{methodSteps to concatenate}

\item{recursive}{Needed for compatibility with generic, otherwise ignored}
}
\description{
concatenate
}

