\docType{methods}
\name{c,originator-method}
\alias{c,originator-method}
\title{concatenate}
\usage{
\S4method{c}{originator}(x, ..., recursive = FALSE)
}
\arguments{
  \item{x,...}{originators to concatenate}

  \item{recursive}{Needed for compatibility with generic,
  otherwise ignored}
}
\description{
concatenate
}

