\docType{methods}
\name{c,references-method}
\alias{c,references-method}
\title{concatenate}
\usage{
\S4method{c}{references}(x, ..., recursive = FALSE)
}
\arguments{
  \item{x,...}{referencess to concatenate}

  \item{recursive}{Needed for compatibility with generic,
  otherwise ignored}
}
\description{
concatenate
}

