% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{methods}
\name{c,responsibleParty-method}
\alias{c,responsibleParty-method}
\title{concatenate}
\usage{
\S4method{c}{responsibleParty}(x, ..., recursive = FALSE)
}
\arguments{
\item{x,...}{responsiblePartys to concatenate}

\item{recursive}{Needed for compatibility with generic, otherwise ignored}
}
\description{
concatenate
}

