% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{class-stmml:unit}
\alias{class-stmml:unit}
\title{concatenate}
\usage{
\S4method{c}{`stmml:unit`}(x, ..., recursive = FALSE)
}
\arguments{
\item{x,...}{units to concatenate}

\item{recursive}{Needed for compatibility with generic, otherwise ignored}
}
\description{
concatenate
}

