% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{class-stmml:unitType}
\alias{class-stmml:unitType}
\title{concatenate}
\usage{
\S4method{c}{`stmml:unitType`}(x, ..., recursive = FALSE)
}
\arguments{
\item{x,...}{objects to concatenate}

\item{recursive}{Needed for compatibility with generic, otherwise ignored}
}
\description{
concatenate
}

