% Generated by roxygen2 (4.0.0): do not edit by hand
\name{coverage}
\alias{coverage}
\title{get coverage information}
\usage{
coverage(coverage)
}
\arguments{
\item{coverage}{an object from which we extract coverage information}
}
\value{
content of a coverage node
}
\description{
get coverage information
}

