\docType{data}
\name{eml}
\alias{eml}
\title{Default and current chunk options}
\format{List of 4
 $ get    :function (name, default = FALSE)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 6 9 11 3 9 3 6 11
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x36fded8> 
 $ set    :function (...)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 12 9 19 3 9 3 12 19
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x36fded8> 
 $ merge  :function (values)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 20 11 20 55 11 55 20 20
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x36fded8> 
 $ restore:function (target = value)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 21 13 21 56 13 56 21 21
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x36fded8>}
\usage{
  eml
}
\description{
  Default and current chunk options
}
\keyword{datasets}

