% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml.R
\name{eml}
\alias{eml}
\title{eml constructor function}
\usage{
eml(dat = NULL, title = "EML_metadata", creator = NULL, contact = NULL,
  coverage = eml_coverage(scientific_names = NULL, dates = NULL,
  geographic_description = NULL, NSEWbox = NULL), methods = new("methods"),
  col.defs = NULL, unit.defs = NULL, col.classes = NULL,
  custom_units = NULL, custom_types = NULL, ...,
  additionalMetadata = new("ListOfadditionalMetadata"), citation = NULL,
  software = NULL, protocol = NULL)
}
\arguments{
\item{dat}{a data.frame object or path to csv}

\item{title}{for the metadata.  Also used as the csv filename.}

\item{creator}{a ListOfcreator object, character string, or person object.
Otherwise loaded from value set by eml_config.}

\item{contact}{a contact object, character string, or person object.
Otherwise loaded from value set by eml_config.}

\item{coverage}{a coverage object, such as created by the eml_coverage constructor function. (optional)}

\item{methods}{a method object or plain text string, documenting additional methods}

\item{col.defs}{Natural language definitions of each column. Should be a character
vector of length equal to each column, defined in the same order as the columns are given.}

\item{unit.defs}{A list of length equal to the number of columns defining the units for each column. See examples.}

\item{col.classes}{column classes, primarily for use if dat is 
a path to a csv rather than a native R object. Optional otherwise.}

\item{custom_units}{a list of custom units. Leave as NULL and these will be automatically
 detected from the EMLConfig after a user declares any custom units using
\code{\link{eml_define_unit}}.  Advanced users can alternatively just give a list
of custom_unit definitions here.  See \code{\link{eml_define_unit}} for details.}

\item{custom_types}{custom unit types, from \code{\link{eml_define_unitType}}}

\item{...}{additional slots passed to the dataset constructor `new("dataset", ...)`}

\item{additionalMetadata}{an additionalMetadata object}

\item{citation}{a citation object (optional, an EML file can have either 
a dataset OR citation OR software OR protocol as its top-level object)}

\item{software}{a software object (optional)}

\item{protocol}{a protocol object (optional)}
}
\description{
eml constructor function
}

