% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml_attributeList.R
\name{eml_attributeList}
\alias{eml_attributeList}
\title{eml_attributeList}
\usage{
eml_attributeList(dat = NULL, col.defs = NULL, unit.defs = NULL,
  col.classes = NULL, col.names = NULL)
}
\arguments{
\item{dat}{a data.frame (or a data.set which has col.defs and unit.defs)}

\item{col.defs}{the column definitions (if not given in meta)}

\item{unit.defs}{the unit definitions (if not given in meta)}

\item{col.classes}{column classes, primarily for use if dat is 
a path to a csv rather than a native R object. Optional otherwise.}

\item{col.names}{names of the columns, if dat is a csv file path.}
}
\description{
eml_attributeList
}

