% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml_config.R
\name{eml_get_config}
\alias{eml_get_config}
\title{eml get config}
\usage{
eml_get_config(x = ls(envir = EMLConfig), default = NULL)
}
\arguments{
\item{x}{name of the configuration option desired (gives all by default)}

\item{default}{default value to use if not found}
}
\value{
display the current configuration options
}
\description{
eml get config
}

