% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml_publish.R
\name{eml_publish}
\alias{eml_publish}
\title{Publish EML}
\usage{
eml_publish(file, ..., destination = c("figshare", "knb"), validate = TRUE,
  schema_only = FALSE)
}
\arguments{
\item{file}{The local path of an EML file}

\item{...}{Additional metadata options.  If publishing to Figshare, 
these correspond to the options in \code{fs_new_article}. See .}

\item{destination}{the publishing destination, e.g. "figshare"}

\item{validate}{logical, should we run validation first? Default true}

\item{schema_only}{logical, option for the validator function. Check only against 
the schema directly? see \code{\link{eml_validate}} for details.}
}
\value{
Uploads the dataset and file to destination specified returns an id
}
\description{
Push EML and data files to an online server and update EML Metadata accordingly
}
\examples{
\dontrun{
   file <- system.file("doc", "my_eml_data.xml", package="EML")
   id <- eml_publish(file, description="Example EML file from EML",
                     categories = "Ecology", tags = "EML")
}
}

