% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml_read.R
\name{eml_read}
\alias{eml_read}
\alias{read.eml}
\title{Read EML files}
\usage{
eml_read(file, ...)
}
\arguments{
\item{file}{an external eml (xml) file, or XMLInternalDocument}

\item{...}{optional additional arguments to xmlParse}
}
\value{
an EML object
}
\description{
Read EML files
}

