\name{eml_read}
\alias{eml_read}
\alias{read.eml}
\title{read eml files (.xml)}
\usage{
  eml_read(file,
    get = c("eml", "data.frame", "attributeList"), ...)
}
\arguments{
  \item{file}{the name or path to the file (see ?xmlParse
  for details)}

  \item{get}{the name of the object that should be
  returned}
}
\description{
  read eml files (.xml)
}

