% Generated by roxygen2 (4.0.1): do not edit by hand
\name{eml_read}
\alias{eml_read}
\alias{read.eml}
\title{Read EML files}
\usage{
eml_read(file, ...)
}
\arguments{
\item{file}{an external eml (xml) file, or XMLInternalDocument}

\item{...}{optional additional arguments to xmlParse}
}
\value{
an EML object
}
\description{
Read EML files
}

