% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coverage.R
\name{get_taxonomicCoverage}
\alias{get_taxonomicCoverage}
\title{get_taxonomicCoverage}
\usage{
get_taxonomicCoverage(taxonomicCoverage)
}
\arguments{
\item{taxonomicCoverage}{a taxonomicCoverage class object}
}
\value{
a data frame with rank names as column names
}
\description{
get_taxonomicCoverage
}
\details{
extract information from a taxonomicCoverage object and construct a data frame
}
\examples{
f <- system.file("xsd/test/eml-i18n.xml", package="EML")
eml <- read_eml(f)
df <- get_taxonomicCoverage(eml@dataset@coverage@taxonomicCoverage)
}

