% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_eml.R
\name{read_eml}
\alias{read_eml}
\title{read_eml}
\usage{
read_eml(file, ...)
}
\arguments{
\item{file}{path to an EML file}

\item{...}{additional arguments to \code{\link{xmlParse}}}
}
\value{
an eml object (S4 object)
}
\description{
read_eml
}
\examples{
f <- system.file("xsd/test", "eml.xml", package = "EML")
eml <- read_eml(f)
}

