% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_eml.R
\name{read_eml}
\alias{read_eml}
\title{read_eml}
\usage{
read_eml(x, from = "xml")
}
\arguments{
\item{x}{path to an EML file}

\item{from}{explicit type for the input format. Possible values:
"xml", "json", "list", or "guess" with "xml" as the default.}
}
\value{
an emld object (list / S3 object)
}
\description{
Read an EML file into R as an emld object.
}
\examples{
f <- system.file("extdata", "example.xml", package = "emld")
eml <- read_eml(f)
}
