% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_party.R
\name{set_creator}
\alias{set_creator}
\title{Create a creator object}
\usage{
set_creator(givenNames, surName, organizationNames, positionNames, addresses,
  phones, electronicMailAddresses, onlineUrls, userIds, salutations)
}
\arguments{
\item{givenNames}{(character) One or more given names for the party's individualName}

\item{surName}{(character) The sur name of the party's individualName}

\item{organizationNames}{(character) One or more organization names for the party}

\item{positionNames}{(character) One or more position names for the party}

\item{addresses}{(address) One or more addresses for the party}

\item{phones}{(character|phone) One or more phones for the party}

\item{electronicMailAddresses}{(character) One or more email addresses for the party}

\item{onlineUrls}{(character) One or more online URLs for the party}

\item{userIds}{(userId) One or more user IDs for the party}

\item{salutations}{(character) One or more salutations for the party's individualName}
}
\value{
(creator) A creator object
}
\description{
Create a creator object
}
\examples{
# A simple example
set_creator("Bryce", "Mecum")
# A more complete example
set_creator("Bryce",
            "Mecum",
            "National Center for Ecological Analysis and Synthesis",
            "Scientific Software Engineer",
            set_address(c("735 State St.", "Suite 300"),
                        "Santa Barbara",
                        "California",
                        "93101-5504",
                        "USA"),
            set_phone("1-555-555-5555"),
            "mecum@nceas.ucsb.edu",
            "https://brycemecum.com",
            set_userId("https://orcid.org/0000-0002-0381-3766",
                       "https://orcid.org"))
}
