% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_section.R
\name{set_section}
\alias{set_section}
\title{set_section}
\usage{
set_section(file = NULL)
}
\arguments{
\item{file}{path to a file providing formatted input text, see details.}
}
\value{
a ListOfsection object that can be used in constructing any element with section slots, see examples
}
\description{
For any EML element that takes a "section" argument, this function can be used to generate the appropriate EML from a markdown-formatted file.
}
\details{
If the `rmarkdown` package is installed, then the input file can 
be a Microsoft Word (.docx) file, a markdown file, or other file
recognized by Pandoc (see http://pandoc.org), which will automate the conversion
to a docbook. Otherwise, the input file should already be in docbook format (with 
.xml or .dbk extension).  Note that pandoc comes pre-installed in RStudio and is 
required for the rmarkdown package.
}
\examples{
f <- system.file("examples/hf205-methods.md", package = "EML")
set_section(f)
new("description", section = set_section(f))

## Can also import from methods written in a .docx MS Word file. 
f <- system.file("examples/hf205-methods.docx", package = "EML")
set_section(f)
new("description", section = set_section(f))

}

