\docType{methods}
\name{show,coverage-method}
\alias{show,coverage-method}
\title{Show coverage pretty-print format using yaml representation}
\usage{
\S4method{show}{coverage}(object)
}
\arguments{
  \item{object}{an EML coverage object}
}
\value{
pretty print coverage in yaml format
}
\description{
Show coverage pretty-print format using yaml representation
}

