% Generated by roxygen2 (4.0.2): do not edit by hand
\name{species}
\alias{species}
\title{get species information}
\usage{
species(taxonomicCoverage)
}
\arguments{
\item{taxonomicCoverage}{an object from which taxonomic coverage information might be extracted}
}
\value{
species names as a character string
}
\description{
get species information
}

