% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_eml.R
\name{write_eml}
\alias{write_eml}
\title{write_eml}
\usage{
write_eml(eml, file, namespaces = NULL, ns = "eml", ...)
}
\arguments{
\item{eml}{an emld class object}

\item{file}{file name to write XML.}

\item{namespaces}{named character vector of additional XML namespaces to use.}

\item{ns}{root namespace abbreviation}

\item{...}{additional arguments to \code{\link{write_xml}}}
}
\value{
If file is not specified, the result is a character string containing
   the resulting XML content. Otherwise return silently.
}
\description{
write_eml
}
\examples{
f <- system.file("extdata", "example.xml", package = "emld")
eml <- read_eml(f)
write_eml(eml, "test.xml")
eml_validate("test.xml")
unlink("test.xml") # clean up
}
